requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==3.0.10
python-dotenv==1.0.0
Flask-Bcrypt==1.0.1
mysql-connector-python==8.1.0
Werkzeug==2.3.6
itsdangerous==2.1.2